Added weight field Some fixes

+++ DataGenerator/src/main/java/org/generator/model/data/Word.java
@@ -4,14 +4,14 @@ import java.util.List;
 
 public class Word {
     private String value;
-    private Integer weight;
+    private Double weight;
     private List<Word> next;
 
     public String getValue() {return value;}
     public void setValue(String value) {this.value = value;}
 
-    public Integer getWeight() {return weight;}
-    public void setWeight(Integer weight) {this.weight = weight;}
+    public Double getWeight() {return weight;}
+    public void setWeight(Double weight) {this.weight = weight;}
 
     public List<Word> getNext() {return next;}
     public void setNext(List<Word> next) {this.next = next;}
